summary tab

+++ src/org/urbanjaguar/antennaanalyzer/FreqSWR.java
@@ -6,14 +6,14 @@ import java.util.Scanner;
  * Created by chris on 6/8/19.
  */
 public class FreqSWR {
-    private Double frequency;
-    private Double VSWR;
+    private Float frequency;
+    private Float VSWR;
 
     public String toString() {
         return frequency.toString() + "|" + VSWR.toString();
     }
 
-    public FreqSWR (double frequency, double VSWR) {
+    public FreqSWR (float frequency, float VSWR) {
         this.frequency = frequency;
         this.VSWR = VSWR;
     }
@@ -21,16 +21,16 @@ public class FreqSWR {
     public FreqSWR (String rawData) {
         Scanner getData = new Scanner(rawData);
         getData.useDelimiter("\\|");
-        this.frequency = getData.nextLong()/1000000.0d;
-        this.VSWR = getData.nextInt()/1000.0d;
+        this.frequency = getData.nextLong()/1000000.0f;
+        this.VSWR = getData.nextInt()/1000.0f;
         getData.close();
     }
 
-    public double getFrequency() {
+    public float getFrequency() {
         return frequency;
     }
 
-    public double getVSWR() {
+    public float getVSWR() {
         return VSWR;
     }
 }
